First step of re-organizing server-side code. Factoring common server infrastructure into a separate subproject, so that individual server modules can be moved to their own subprojects.
