+++ server-env/build.gradle.kts
@@ -1,9 +1,8 @@
 plugins {
     alias(libs.plugins.gretty)
     alias(libs.plugins.kotlinSerialization)
-    id("war")
     id("java-library")
-    id("org.jetbrains.kotlin.jvm")
+    alias(libs.plugins.jetbrains.kotlin.jvm)
     alias(libs.plugins.ksp)
 }
 
@@ -18,17 +17,10 @@ java {
 
 dependencies {
     ksp(project(":processor"))
+    implementation(project(":processor-annotations"))
     implementation(project(":identity"))
     implementation(project(":identity-flow"))
-    implementation(project(":processor-annotations"))
-    implementation(project(":identity-issuance"))
-    implementation(project(":identity-csa"))
-    implementation(project(":identity-mdoc"))
-    implementation(project(":identity-sdjwt"))
-    implementation(project(":identity-doctypes"))
-    implementation(project(":server-env"))
 
-    implementation(libs.javax.servlet.api)
     implementation(libs.kotlinx.datetime)
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.kotlinx.coroutines.core)
@@ -38,9 +30,6 @@ dependencies {
     implementation(libs.mysql)
     implementation(libs.ktor.client.core)
     implementation(libs.ktor.client.java)
-    implementation(libs.nimbus.oauth2.oidc.sdk)
 
     testImplementation(libs.junit)
 }
-
-gretty {}
