+++ server-env/src/main/java/com/android/identity/server/ServletConfiguration.kt
@@ -1,16 +1,11 @@
-package com.android.identity.wallet.server
+package com.android.identity.server
 
 import com.android.identity.flow.server.Configuration
 import jakarta.servlet.ServletConfig
 
-class ServerConfiguration(private val servletConfig: ServletConfig) : Configuration {
-
+internal class ServletConfiguration(private val servletConfig: ServletConfig) : Configuration {
     override fun getValue(key: String): String? {
         val value = servletConfig.getInitParameter(key)
         return value
-    }
-
-    companion object {
-        private const val TAG = "ServerConfiguration"
     }
 }
