+++ server/src/main/resources/resources/common_configuration.json
@@ -1,0 +1,2 @@
+{
+}
