+++ server/src/main/resources/resources/www/login.html
@@ -5,7 +5,7 @@
 </head>
 <body>
 <h1>Wallet server admin login</h1>
-<form method="post" action="admin/login" style="display: inline-block; text-align: right">
+<form method="post" action="login" style="display: inline-block; text-align: right">
     <label>
         Admin password:
         <input type="password" name="password"/>
