+++ server/src/main/resources/resources/www/password.html
@@ -5,7 +5,7 @@
 </head>
 <body>
 <h1>Wallet server change admin password</h1>
-<form method="post" action="admin/password" style="display: inline-block; text-align: right">
+<form method="post" action="password" style="display: inline-block; text-align: right">
     <label>
         Old password:
         <input type="password" name="oldPassword"/>
