+++ server/src/main/resources/resources/www/update.html
@@ -5,7 +5,7 @@
     </head>
     <body>
         <h1>Update a document</h1>
-        <form method="post" action="admin/updateDocument" style="display: inline-block; text-align: right">
+        <form method="post" action="updateDocument" style="display: inline-block; text-align: right">
             <label>
                 Client Id:
                 <input type="text" name="clientId"/>
