+++ settings.gradle.kts
@@ -59,3 +59,4 @@ include(":wallet")
 include(":server")
 include(":appverifier")
 include(":appholder")
+include(":server-env")
